Terminate instance after 5400"

+++ run.sh
@@ -118,7 +118,12 @@ for VAL in $RES; do
   JOB_ID=$(echo "${TAG}" | awk -F '-' '{ print $NF }')
   JOB_STATUS=$(curl -H "authorization: token ${GH_PAT}" "https://api.github.com/repos/${REPO}/actions/jobs/${JOB_ID}" | jq -cr '.status')
   if [ "${JOB_STATUS}" != "queued" ] && [ "${JOB_STATUS}" != "in_progress" ]; then
-    aws ec2 terminate-instances --instance-ids "${ID}"
+    EC2_LAUNCH_TIME=$(aws ec2 describe-instances --instance-ids "${ID}" --query 'Reservations[*].Instances[*].LaunchTime' --output text)
+    LAUNCH_TIME_EPOCH=$(date -d $EC2_LAUNCH_TIME +"%s"
+    diff=$(expr $CURRENT_TIME_EPOCH - $LAUNCH_TIME_EPOCH)
+    if [ $diff -gt 5400 ]; then
+      aws ec2 terminate-instances --instance-ids "${ID}"
+    fi
   fi
 done
